move polar polygon logic to polar/helper.js
- add move findIndexOfMin to lib/search.js
