+++ src/lib/search.js
@@ -11,6 +11,7 @@
 
 var isNumeric = require('fast-isnumeric');
 var loggers = require('./loggers');
+var identity = require('./identity');
 
 // don't trust floating point equality - fraction of bin size to call
 // "on the line" and ensure that they go the right way specified by
@@ -112,4 +113,27 @@ exports.roundUp = function(val, arrayIn, reverse) {
         else high = mid - dhigh;
     }
     return arrayIn[low];
+};
+
+/* find index in array 'arr' that minimizes 'fn'
+ *
+ * @param {array} arr : array where to search
+ * @param {fn (optional)} fn : function to minimize,
+ *   if not given, fn is the identity function
+ * @return {integer}
+ */
+exports.findIndexOfMin = function(arr, fn) {
+    fn = fn || identity;
+
+    var min = Infinity;
+    var ind;
+
+    for(var i = 0; i < arr.length; i++) {
+        var v = fn(arr[i]);
+        if(v < min) {
+            min = v;
+            ind = i;
+        }
+    }
+    return ind;
 };
